requirements.txt
Django==5.0.6
django-debug-toolbar==4.3.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
pillow==10.3.0
psycopg2-binary==2.9.9
sqlparse==0.5.0
djoser==2.2.3
requests==2.32.3
